// src/modules/estudante/dto/estudante-integracao.dto.ts
export class EstudanteIntegracaoDto {
  id: string;
  inep: number;
  idpes: number;
  cod_matricula: string;
  cod_turma: number;
  nome: string;
  nome_mae: string;
  nome_pai: string;
  nome_resp: string;
  log_no: string;
  complemento: string;
  bai_no: string;
  loc_no: string;
  ufe_sg: string;
  cep: string;
  data_nasc: Date;
}